Split files into packages
