+++ cli/add.go
@@ -1,10 +1,11 @@
-package main
+package cli
 
 import (
 	"fmt"
 	"os"
 
 	"github.com/99designs/aws-vault/prompt"
+	"github.com/99designs/aws-vault/vault"
 	"github.com/99designs/keyring"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"gopkg.in/alecthomas/kingpin.v2"
@@ -14,6 +15,24 @@ type AddCommandInput struct {
 	Profile string
 	Keyring keyring.Keyring
 	FromEnv bool
+}
+
+func ConfigureAddCommand(app *kingpin.Application) {
+	input := AddCommandInput{}
+
+	cmd := app.Command("add", "Adds credentials, prompts if none provided")
+	cmd.Arg("profile", "Name of the profile").
+		Required().
+		StringVar(&input.Profile)
+
+	cmd.Flag("env", "Read the credentials from the environment").
+		BoolVar(&input.FromEnv)
+
+	cmd.Action(func(c *kingpin.ParseContext) error {
+		input.Keyring = keyringImpl
+		AddCommand(app, input)
+		return nil
+	})
 }
 
 func AddCommand(app *kingpin.Application, input AddCommandInput) {
@@ -41,7 +60,7 @@ func AddCommand(app *kingpin.Application, input AddCommandInput) {
 	}
 
 	creds := credentials.Value{AccessKeyID: accessKeyId, SecretAccessKey: secretKey}
-	provider := &KeyringProvider{Keyring: input.Keyring, Profile: input.Profile}
+	provider := &vault.KeyringProvider{Keyring: input.Keyring, Profile: input.Profile}
 
 	if err := provider.Store(creds); err != nil {
 		app.Fatalf(err.Error())
@@ -56,7 +75,7 @@ func AddCommand(app *kingpin.Application, input AddCommandInput) {
 		return
 	}
 
-	sessions, err := NewKeyringSessions(input.Keyring, profiles)
+	sessions, err := vault.NewKeyringSessions(input.Keyring, profiles)
 	if err != nil {
 		app.Fatalf(err.Error())
 		return
