+++ server/alias_linux.go
@@ -1,6 +1,6 @@
 // build: linux
 
-package main
+package server
 
 import "os/exec"
