+++ server/alias_windows.go
@@ -1,6 +1,6 @@
 // build: windows
 
-package main
+package server
 
 import "fmt"
